Giant update, add btv SFs code, compiled FTFunctions.cpp with environment flag FOURTOPNAOD_COMPILE, separated out FTFunctions.cpp, etc.
